move everything inside Discoverer
